Use jest for testing
